Fix alignment of structs plus do more to pose.hpp
Fix the "#pragma pack" directives. The directive needed a byte alignment
value. In our case we want to align on 1 byte boundaries. We also want
to restore the original byte alignment setting. So I used the push and
pop feature of #pragma pack. The GCC docs for #pragma pack are here:

https://gcc.gnu.org/onlinedocs/gcc-4.9.3/gcc/Structure-Packing-Pragmas.html#Structure-Packing-Pragmas

Add a message_coutner field to PoseMessage. This will allow us to reject
old messages that arrive out of order.

Finally, clean up pose.hpp. Use the C/C++ idiom to define a struct with
a typedef. This prevents an object, function, or enumerator from hiding
the struct name. The problem with the old code is subtle and it's
described well by the following stackoverflow answers:

    This describes what's going on:
    https://stackoverflow.com/a/1675446

    This describes why it matters:
    https://stackoverflow.com/a/612476

See GitHub issue: #14

Co-authored-by: Jason Brauer <JasonBrauer@users.noreply.github.com>