Fix endian functions not building on macOS
On macOS the endian-changer functions like htonll are built-in and the compiler doesn't let you re-declare them. This commit moves our htonll and ntohll functions to a namespace.

Also update the udp_sender and udp_receiver code to use the namespaced version of htonll. While in some ways it would be better to let macOS use the built-in functions, it would also require us to selectively declare htonll and ntohll. It's probably easier to maintain the code if we avoid compiler directives like #if as those require special attention and testing.

#14