Allow hostnames for airsim
Before this change the AirSim host needed to be an IP address. After this change the AirSim host could be a hostname string. 

This will help us if in the future we're able to run AirSim with docker. We'll be able to use the docker host name to send packets to right. place.